Refactor aspects BackEndApi
findById issue solved through adding 'this' before axios.crudMethod on BackEndApi. This way, this.backEndApi.findCardById stopped being undefined on Home.

Changed BackEndApi code to be shown in one line. Also, removed everything before '/api' on the URL.

Changed class name to be BackEndApi, instead of CallApi, which was thus  named due to lack of attention